creating relation between course and coupon entities

+++ src/main/java/com/example/couponapi/repository/CouponRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface CouponRepository extends JpaRepository<Coupon, Integer> {
+public interface CouponRepository extends JpaRepository<Coupon, Long> {
     boolean existsByCode(String code);
     Optional<Coupon> findByCode(String code);
 }
